0.1.0 preview Release
With the 0.1.0 release, the Changelog is now required to be updated throughout the development of this package. Let's go over some highlights:

ArraySetting<T> added. This allows settings to be picked much like an Enum, but from an array.

FloatSettingInterpreter and IntSettingInterpreter can now take a Min and Max value that they will use to reinterpret the Slider (or Scrollbar) value. This allows for any Slider of x and y to always be reinterpreted as a value from Min to Max.

forceView has been renamed to forceUI, as forceView was confusing Julius about its use (he thought it meant the value wasn't updated if forceView was off).

InputFieldInterpreters now have a method to validate input as well.

Settings added for:
- QualityLevel
- Resolution
- Anti Aliasing
- Shadow Distance
- Shadow Resolution

And a few settings have gotten Helpboxes in their editors (notebly Screen Settings for their incomplience in the Editor).