Setting Undo functionality and refactoring
Notibly, FloatSetting and IntSetting have a min and max value now that can be overriden. If both of these are set, the inspector reflects this with a slider.

EnumSetting now has value as its T value and enumValue as its Enum value.

The rest of the changes in this commit are automatic changes based on refactoring.